requirements.txt
speechrecognition
pyttsx3
google-generativeai
requests
pocketsphinx
psutil
streamlit
yt-dlp
ollama
beautifulsoup4
pywhatkit
python-telegram-bot
dateparser
pandas
pyautogui